// src/api/generated/model/cartItemDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with OpenIddict Authentication
 * OpenAPI spec version: v1
 */

export interface CartItemDto {
  id?: string;
  productId?: string;
  /** @nullable */
  productName?: string | null;
  unitPrice?: number;
  quantity?: number;
  totalPrice?: number;
}